+++ src/components/util/storybook.js
@@ -57,3 +57,44 @@ export const StoryItem = ({ children }) => (
 StoryItem.propTypes = {
   children: PropTypes.node,
 };
+
+/**
+ * For displaying important info in MDX docs
+ */
+export const Info = ({ type, children }) => {
+  const accentMap = {
+    warn: FDS.COLOR_ORANGE,
+    info: FDS.COLOR_SKY,
+  };
+  const emojiMap = {
+    warn: '⚠️',
+    info: '💡',
+  };
+
+  const style = {
+    border: `1px solid ${accentMap[type]}`,
+    borderTop: `8px solid ${accentMap[type]}`,
+    display: 'flex',
+    alignItems: 'center',
+  };
+
+  return (
+    <blockquote
+      style={style}
+      className="margin--bottom padding--top--half padding--bottom--half padding--left padding--right rounded--all"
+    >
+      <span className="padding--right fontSize--2xl">{emojiMap[type]}</span>
+      <span className="color--heading">{children}</span>
+    </blockquote>
+  );
+};
+
+Info.defaultProps = {
+  type: 'info',
+};
+
+Info.propTypes = {
+  children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.node), PropTypes.node])
+    .isRequired,
+  type: PropTypes.oneOf(['warn', 'info']),
+};